Align the longer translated "No data" text properly
While testing my translation on a small screen of Jolla 1 I noticed that the Russian string for "No data - Click to fetch data" is longer and does not fit the width of the graph area. I had to make it 2 lines, but it looks much nicer if the text is centered. There is no visible change in the English UI.